Compares two cell survival curves using ANOVA.

It helps compare two cell survival curves using the cellsurvLQdiff function of the CFAssay package, without having the user go through the data wrangling steps necessary for that. Visit https://bioconductor.org/packages/release/bioc/html/CFAssay.html for more details about the method.

Usage

compareCurves(data, cline1, cline2, method = "ml", PEmethod = "fit")

Arguments

data

A data frame containing at least the following five columns: "cline", "Exp", "dose", "ncells", "ncolonies".

cline1

Name of the cell-line/group that you want to compare.

cline2

Name of another cell-line/group that you want to compare.

method

Method used for the fit. It's "ml" (maximum likelihood) by default. Can be "ls" (least squares) or "franken" (weighted least squares as described by Franken eta al.(2006)).

PEmethod

Controls the value of the plating efficiencies. "fit" calculates fitted plating efficiencies as model parameters, "fix" uses fixed ones calculated from the observed zero dose data.

Value

An object of class cellsurvLQdiff, as returned by cellsurvLQdiff.

Examples

datatab <- CASP8_data
compareCurves(datatab, "shCASP8-N", "shCASP8-B+Z+N")
#> ****** Cell type 1: shCASP8-N   ||   Cell type 2: shCASP8-B+Z+N ******
#> 
#> Overall comparison test for coefficients alpha and beta of LQ-models 
#> ==================================================================== 
#> method = ml 
#> PEmethod = fit 
#> 
#> 6 PEs fitted as intercepts. To look at, use simple R print function. 
#> Null hypothesis (Model 1): one set of shape parameters alpha and beta for all data 
#> ---------------------------------------------------------------------------------- 
#>          Estimate  Std. Error    t value     Pr(>|t|)
#> alpha -0.01910478 0.020810067 -0.9180547 3.722228e-01
#> beta  -0.03692732 0.003798381 -9.7218568 4.061270e-08
#> 
#> Goodness-of-fit values 
#> 	Residual Deviance: 12.85619 
#> 	Total sum of squared weighted residuals rsswTot: 13.19573 
#> 	Residual Degrees of Freedom: 16 
#> 	Dispersion parameter: 0.8247333 
#> 
#> Alternative hypothesis (Model 2): two sets of shape parameters alpha and beta 
#> ----------------------------------------------------------------------------- 
#>                              Estimate  Std. Error    t value     Pr(>|t|)
#> alpha:curvesshCASP8-B+Z+N -0.02224331 0.031717703 -0.7012901 4.946148e-01
#> alpha:curvesshCASP8-N     -0.01613085 0.030528750 -0.5283823 6.055094e-01
#> beta:curvesshCASP8-B+Z+N  -0.03710309 0.005804313 -6.3923312 1.675728e-05
#> beta:curvesshCASP8-N      -0.03678049 0.005559103 -6.6162636 1.158171e-05
#> 
#> Goodness-of-fit values 
#> 	Residual Deviance: 12.60644 
#> 	Total sum of squared weighted residuals rsswTot: 12.89855 
#> 	Residual Degrees of Freedom: 14 
#> 	Dispersion parameter: 0.9213253 
#> 
#> Analysis of Variance Table and F-test
#> Model 2 versus Model 1
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1        16     12.856                          
#> 2        14     12.606  2  0.24975 0.1355 0.8744
#> 
#> *** Analysis by CellSurvAssay v1.0.0 ***
compareCurves(datatab, "shCASP8-N", "shCASP8-B+Z+N", method = "ls", PEmethod = "fix")
#> ****** Cell type 1: shCASP8-N   ||   Cell type 2: shCASP8-B+Z+N ******
#> 
#> Overall comparison test for coefficients alpha and beta of LQ-models 
#> ==================================================================== 
#> method = ls 
#> PEmethod = fix 
#> 
#> Null hypothesis (Model 1): one set of shape parameters alpha and beta for all data 
#> ---------------------------------------------------------------------------------- 
#>          Estimate  Std. Error    t value     Pr(>|t|)
#> alpha -0.01448839 0.019445780  -0.745066 4.670361e-01
#> beta  -0.03843262 0.003674907 -10.458122 1.467656e-08
#> 
#> 
#> Goodness-of-fit values 
#> 	Total sum of squared residuals rssTot: 0.1126081 
#> 	Residual Degrees of Freedom: 16 
#> 	Multiple R-squared: 0.9986025 
#> 
#> Alternative hypothesis (Model 2): two sets of shape parameters alpha and beta 
#> ----------------------------------------------------------------------------- 
#>                               Estimate Std. Error    t value     Pr(>|t|)
#> alpha:curvesshCASP8-B+Z+N -0.020664576 0.02914708 -0.7089759 4.899782e-01
#> alpha:curvesshCASP8-N     -0.008312201 0.02914708 -0.2851812 7.796808e-01
#> beta:curvesshCASP8-B+Z+N  -0.037657377 0.00550828 -6.8365033 8.107448e-06
#> beta:curvesshCASP8-N      -0.039207870 0.00550828 -7.1179873 5.188559e-06
#> 
#> 
#> Goodness-of-fit values 
#> 	Total sum of squared residuals rssTot: 0.1106845 
#> 	Residual Degrees of Freedom: 14 
#> 	Multiple R-squared: 0.9986266 
#> 
#> Analysis of Variance Table and F-test
#> Model 2 versus Model 1
#>   Res.Df     RSS Df Sum of Sq      F Pr(>F)
#> 1     16 0.11261                           
#> 2     14 0.11069  2 0.0019235 0.1216 0.8864
#> 
#> *** Analysis by CellSurvAssay v1.0.0 ***