Fits the Linear Quadratic Model

It helps fit the linear quadratic model using the cellsurvLQfit function of the CFAssay package, but automates the data wrangling steps, thus making it more convenient to use for beginner R users. Visit https://bioconductor.org/packages/release/bioc/html/CFAssay.html for more details about the method of the fit.

Usage

lqmodelFit(data, ctype, method = "ml", PEmethod = "fit")

Arguments

data

A data frame containing at least the following five columns with these exact names: "cline", "Exp", "dose", "ncells", "ncolonies".

ctype

Name of the cell-line/group for which the model is to be fit.

method

Method used for the fit. It's "ml" (maximum likelihood) by default. Also accepts "ls" (least squares) or "franken" (weighted least squares as described by Franken eta al.(2006)).

PEmethod

Controls the value of the plating efficiencies. "fit" calculates fitted plating efficiencies as model parameters, "fix" uses fixed ones calculated from the observed zero dose data.

Value

An object of class cellsurvLQfit, as returned by cellsurvLQfit.

Examples

datatab <- CASP8_data
lqmodelFit(datatab, "control-B")
#> *** Coefficients of LQ-model for cell survival *** 
#> method = ml 
#> PEmethod = fit 
#> 
#> Logarithmic plating efficiencies PE fitted as intercepts 
#> see remark in the manual, 1.2 
#>      Estimate Std. Error   t value     Pr(>|t|)
#> PE1 -1.241244 0.04588891 -27.04888 2.420266e-08
#> PE2 -1.183913 0.04495808 -26.33371 2.914873e-08
#> PE3 -1.252173 0.04607027 -27.17963 2.340599e-08
#> 
#> Shape parameters alpha and beta 
#>          Estimate  Std. Error    t value     Pr(>|t|)
#> alpha -0.01366794 0.034615893 -0.3948458 7.047115e-01
#> beta  -0.06047490 0.006936648 -8.7181736 5.243901e-05
#> 
#> Observed and fitted plating efficiencies (%): 
#>     Experiment   PE PEfitted
#> PE1          1 29.5     28.9
#> PE2          2 31.0     30.6
#> PE3          3 27.0     28.6
#> 
#> Residual Deviance: 6.7948 
#> Total residual sum of weighted squares rsswTot: 6.841747 
#> Residual Degrees of Freedom: 7 
#> Dispersion parameter: 0.9773924 
#> 
#> Fraction rssw of rsswTot per Experiment 
#>   Experiment rssw perCent
#> 1          1 0.65     9.5
#> 2          2 0.82    12.0
#> 3          3 5.37    78.5
lqmodelFit(datatab, "control-B", method = "ls", PEmethod = "fix")
#> *** Coefficients of LQ-model for cell survival *** 
#> method = ls 
#> PEmethod = fix 
#> 
#> Shape parameters alpha and beta 
#>          Estimate  Std. Error    t value     Pr(>|t|)
#> alpha -0.03085093 0.036336074 -0.8490442 4.239383e-01
#> beta  -0.05626118 0.006866872 -8.1931296 7.824695e-05
#> 
#> Total residual sum of squares rssTot: 0.08600878 
#> Residual Degrees of Freedom: 7 
#> Multiple R-squared: 0.9985152 
#> 
#> Fraction rss of rssTot per Experiment 
#>   Experiment  rss perCent
#> 1          1 0.03    31.4
#> 2          2 0.00     5.0
#> 3          3 0.05    63.6